Start using webpack for typescript transpilation
- Use webpack.dev.ts as config name instead of the default to avoid conflicts with the task webpack-browser-dependencies
- Use webpackIgnore directive for dynamic require and where necessary
- To make webpackIgnore wrok set removeComments to false in some tsconfig
- Move connection-manager in domain-manager to avoid a circular dependency
- Remove app/cache.ts since unused
- Use shell option for cp.spawn in TypeScriptTooling extension
- Use TranspilePlugin plugin to transpiles input files into output files individually without bundling together.
  Use this way to easer the transition from directly using tsc.

Rename
