+++ package.json
@@ -96,14 +96,15 @@
   },
   "scripts": {
     "postinstall": "gulp install",
-    "tsc": "tsc --project app && tsc --project src && tsc --project src/filesystem/impls/appshell/node && tsc --project src/extensions/default/TypeScriptTooling/node",
-    "build": "rimraf ./dist && gulp copy-src-dist && gulp build && npm run tsc && tsc --project src/filesystem/impls/appshell/node && tsc --project src/extensions/default/TypeScriptTooling/node",
+    "webpack": "webpack --config webpack.dev.ts",
+    "tsc": "tsc --project src && npm run webpack",
+    "build": "gulp copy-src-dist && gulp build && npm run tsc",
     "build-optimized": "npm run build && gulp optimize",
     "test": "npm run tslint",
     "tslint:app": "tslint -c tslint.json --project app/tsconfig.json -t stylish",
     "tslint:src": "tslint -c src/tslint.json --project src/tsconfig.json -t stylish",
     "tslint": "npm run tslint:app && npm run tslint:src",
-    "dev": "concurrently --kill-others \"gulp watch\" \"tsc --watch --project app\" \"tsc --watch --project src\" \"tsc --watch --project src/filesystem/impls/appshell/node\" \"tsc --watch --project src/extensions/default/TypeScriptTooling/node\"",
+    "dev": "concurrently --kill-others \"gulp watch\" \"tsc --watch --project src\" \"npm run webpack -- --watch\"",
     "start": "electron .",
     "pack": "npm run build-optimized && electron-builder --dir",
     "dist": "npm run build-optimized && electron-builder",
@@ -153,6 +154,7 @@
     "@types/anymatch": "^1.3.1",
     "@types/fs-extra": "^9.0.13",
     "@types/lodash": "^4.14.188",
+    "@types/webpack-node-externals": "^3.0.4",
     "@types/ws": "0.0.40",
     "@types/xml2js": "^0.4.11",
     "@types/yargs": "^17.0.13",
@@ -184,11 +186,15 @@
     "rewire": "^6.0.0",
     "rimraf": "^3.0.2",
     "tar": "^6.1.12",
+    "transpile-webpack-plugin": "^1.1.3",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
     "tslint": "^5.20.1",
     "typescript": "~4.9.4",
     "vscode-languageserver": "5.3.0-next.1",
     "webpack": "^5.93.0",
     "webpack-cli": "^5.1.4",
+    "webpack-node-externals": "^3.0.0",
     "xmldoc": "0.1.2"
   },
   "overrides": {
